# swagger/swagger.yaml
openapi: 3.0.0
info:
  title: FixingApp Backend API
  description: |
    Backend MVP API for connecting employers with workers.
    
    This API provides endpoints for:
    - Authentication and user management
    - Job posting and management
    - Job applications
    - Worker verification
    - Reviews and ratings
    - Complaints handling
    - Admin functions
    
    **Authentication**: Most endpoints require a JWT Bearer token in the Authorization header.
  version: 1.0.0
  contact:
    name: API Support
    email: support@fixingapp.com
servers:
  - url: http://localhost:3000
    description: Local development server
  - url: http://localhost:3000/api
    description: API base path
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from login endpoint
  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
          description: Error message
        message:
          type: string
          description: Detailed error message
      example:
        error: "Not Found"
        message: "Route GET /api/invalid not found"
...
    User:
      type: object
      properties:
        id:
          type: integer
          example: 1
        phone:
          type: string
          example: "0123456789"
        fullName:
          type: string
          example: "John Doe"
        address:
          type: string
          example: "123 Main St"
        role:
          type: string
          enum: [EMPLOYER, WORKER, ADMIN]
          example: "EMPLOYER"
        createdAt:
          type: integer
          format: int64
          example: 1702719041000
    Job:
      type: object
      properties:
        id:
          type: integer
          example: 1
        employerId:
          type: integer
          example: 1
        title:
          type: string
          example: "Fix leaking pipe"
        description:
          type: string
          example: "Need to fix a leaking pipe in kitchen"
        price:
          type: integer
          format: int64
          example: 150000
        address:
          type: string
          example: "123 Main St, City"
        requiredSkill:
          type: string
          example: "Plumbing"
        status:
          type: string
          enum: [CHUA_LAM, DANG_BAN_GIAO, DA_XONG]
          example: "CHUA_LAM"
        acceptedWorkerId:
          type: integer
          nullable: true
          example: null
        createdAt:
          type: integer
          format: int64
          example: 1702719041000
        updatedAt:
          type: integer
          format: int64
          example: 1702719041000
paths:
...
  /api/auth/register-employer:
    post:
      tags:
        - Auth
      summary: Register as employer
      description: Create a new employer account
      operationId: registerEmployer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - phone
                - password
                - fullName
              properties:
                phone:
                  type: string
                  example: "0123456789"
                password:
                  type: string
                  format: password
                  example: "password123"
                fullName:
                  type: string
                  example: "John Employer"
                address:
                  type: string
                  example: "123 Main St"
            example:
              phone: "0123456789"
              password: "password123"
              fullName: "John Employer"
              address: "123 Main St"
      responses:
        '201':
          description: Employer registered successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  token:
                    type: string
                    example: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
                  user:
                    $ref: '#/components/schemas/User'
        '400':
          description: Bad request (e.g., duplicate phone)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/auth/register-worker:
    post:
      tags:
        - Auth
      summary: Register as worker
      description: Create a new worker account
      operationId: registerWorker
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - phone
                - password
                - fullName
                - skill
              properties:
                phone:
                  type: string
                  example: "0987654321"
                password:
                  type: string
                  format: password
                  example: "password123"
                fullName:
                  type: string
                  example: "Jane Worker"
                address:
                  type: string
                  example: "789 Worker St"
                skill:
                  type: string
                  example: "Plumbing"
            example:
              phone: "0987654321"
              password: "password123"
              fullName: "Jane Worker"
              address: "789 Worker St"
              skill: "Plumbing"
      responses:
        '201':
          description: Worker registered successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  token:
                    type: string
                    example: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
                  user:
                    $ref: '#/components/schemas/User'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /api/jobs/{jobId}/apply:
    post:
      tags:
        - Applications
      summary: Apply to a job
      description: Submit an application to a job (Worker only)
      operationId: applyToJob
      security:
        - bearerAuth: []
      parameters:
        - name: jobId
          in: path
          required: true
          schema:
            type: integer
          example: 1
      responses:
        '201':
          description: Application submitted successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: integer
                  jobId:
                    type: integer
                  workerId:
                    type: integer
                  status:
                    type: string
                    enum: [APPLIED]
        '400':
          description: Bad request (e.g., job already completed, duplicate application)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/jobs/{jobId}/applications:
    get:
      tags:
        - Applications
      summary: Get job applications
      description: Get all applications for a job (Job owner only)
      operationId: getJobApplications
      security:
        - bearerAuth: []
      parameters:
        - name: jobId
          in: path
          required: true
          schema:
            type: integer
          example: 1
      responses:
        '200':
          description: List of applications
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    id:
                      type: integer
                    workerId:
                      type: integer
                    worker:
                      type: object
                    status:
                      type: string
                      enum: [APPLIED, ACCEPTED, REJECTED]
        '403':
          description: Forbidden (not job owner)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/jobs/{jobId}/accept/{workerId}:
    post:
      tags:
        - Applications
      summary: Accept a worker
      description: Accept a worker's application (Job owner only)
      operationId: acceptWorker
      security:
        - bearerAuth: []
      parameters:
        - name: jobId
          in: path
          required: true
          schema:
            type: integer
          example: 1
        - name: workerId
          in: path
          required: true
          schema:
            type: integer
          example: 2
      responses:
        '200':
          description: Worker accepted successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Job'
        '400':
          description: Bad request (e.g., worker didn't apply, already accepted another worker)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Forbidden (not job owner)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/jobs/{jobId}/reject/{workerId}:
    post:
      tags:
        - Applications
      summary: Reject a worker
      description: Reject a worker's application (Job owner only)
      operationId: rejectWorker
      security:
        - bearerAuth: []
      parameters:
        - name: jobId
          in: path
          required: true
          schema:
            type: integer
          example: 1
        - name: workerId
          in: path
          required: true
          schema:
            type: integer
          example: 2
      responses:
        '200':
          description: Worker rejected successfully
        '403':
          description: Forbidden (not job owner)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...